Update library imports related to variable refactor
