+++ pyaerocom/__init__.py
@@ -51,7 +51,7 @@ from .colocation_auto import ColocationSetup, Colocator
 from .tstype import TsType
 from .time_resampler import TimeResampler
 from .io.helpers import search_data_dir_aerocom
-from .variable import get_variable
+from .variable_helpers import get_variable
 from .utils import create_varinfo_table
 from .testdata_access import initialise as initialise_testdata
