+++ pyaerocom/io/aux_read_cubes.py
@@ -188,7 +188,7 @@ def compute_angstrom_coeff_cubes(cube1, cube2, lambda1=None, lambda2=None):
     Cube
         Cube containing Angstrom exponent(s)
     """
-    from pyaerocom.variable import VarNameInfo
+    from pyaerocom.varnameinfo import VarNameInfo
     from cf_units import Unit
     cube1, cube2 = _check_input_iscube(cube1, cube2)
     if lambda1 is None:
